Provide settings per content source to allow for differentiations between them.
The reindex operations of different content sources vary in time, with some being extremely quick (tags & sections) to some taking much longer. Therefore it does not make sense for applications that reindex quickly to implement the ability to cancel a reindex or to allow a reindex for a specific time period necessarily. This pull request provides the flexibility for this and reflects it in the UI.