Use index as key
which is somewhat frowned upon, but this state is local to our component and so we should be fine. The alternative is to use the header key and value, which rerenders the component each time the input chagnes, losing focus; or adding an id property to the header data, which seems excessive.