ci: Update triggers
The commenting behaviour of https://github.com/guardian/actions-riff-raff
inspects the calling payload and tries to associate it with a pull request.

If the build triggered on push completes before the PR is opened,
a comment is not produced. Any subsequent commits to a PR will yield a comment.
This is a little confusing.

This should fix that. By triggering the workflow on `pull_request`
we know that `guardian/actions-riff-raff` will be able to locate a PR to comment on.